New Buttons and CSS Style
Added a new row of buttons at the top of the calculator - making it the first row of buttons.  It is below the number output and above the “AC, +/-, %, /“ buttons.

Created a new CSS class with a different background color for these new buttons. Buttons are the same size as the others and the text is bold on the buttons